Warn fixes

+++ client/engine/src/Layers/LayerAvailRooms.cpp
@@ -49,7 +49,7 @@ namespace LM {
         ImGui::SetNextWindowPos(ImVec2(io.DisplaySize.x * 0.5f, io.DisplaySize.y * 0.5f), ImGuiCond_Always, ImVec2(0.5f, 0.5f));
         if (ImGui::Begin("Rooms", 0, ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_AlwaysAutoResize)) {
             for (size_t i = 0; i < m_Rooms.size(); ++i) {
-                ImGui::PushID(m_Rooms[i].getId());
+                ImGui::PushID(static_cast<int>(m_Rooms[i].getId()));
                 ImGui::Text("RoomId: %10llu", m_Rooms[i].getId());
                 ImGui::SameLine();
                 if (ImGui::Button("Connect")) {
